+++ CHANGELOG.md
@@ -25,6 +25,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Changed
+
+- Clarify `language` column in tooltip (#141).
+
 ### Added
 
 - [REVERTED] Code quality metrics in dashboard (reliability, security, maintainability) (#63).